Add a script to generate Matroska tags XML
We use a tsv file to store the metadata to be set in the video
files. If we are to use mkvpropedit (1) to add tags to the mkv / webm
file, we need to save them in an XML file matching the Matroska tag
format specification (2).

Using mkvpropedit after encoding (instead of using ffmpeg -metadata)
has one advantage: it allows us to use Matroska's notion of tag target
types: a video of a single talk ("chapter") can store metadata about
the full conference video ("movie") it comes from -- in our case that
would apply to the title (conference and talk). We use target values
of 50 ("movie") and 30 ("chapter") because VLC fixates on these
values, mpv is more flexible.

Note that although Matroska specification (2) says: "Multiple items
SHOULD never be stored as a list in a single TagString. If there is
more than one tag of a certain type to be stored, then more than one
SimpleTag SHOULD be used.", we deviate from this recommendation in the
ARTIST tag. We create only one ARTIST tag, even when there are
multiple speakers (in which case they are comma-separated). This is
because players seem to ignore multiple tags: VLC shows only the last,
and GNOME Video shows only the first.

(1) https://mkvtoolnix.download/doc/mkvpropedit.html
(2) https://matroska.org/technical/tagging.html