Add a bash script for rendering a single video
This script contains invocations of ffmpeg and mkvpropedit. We bundle
them together because there will be little point in datalad-saving the
file twice (first without and then with metadata). The inputs are:
collection directory, clip number (= line number of clips.tsv).

The video output format is set to webm, with AV1 and opus codecs. Webm
should be readily playable in browsers as well as in video players. We
use SVT-AV1 preset 6 as it should provide "balance of efficiency and
reasonable compute time" (1). We clip the number of threads (lp) to 1
because we want to schedule jobs requesting 1 CPU. Although YouTube
video already has opus audio, we reencode with lower settings (in
particular, single channel), which are still very reasonable for a
talk recording, to save further megabytes. Taken together, the
settings used should produce about 105 MB for a 25 minute
talk. Single-thread encoding time for 25 min video is 3 hours with
single thread (and 35 mins for 8 threads).

(1) https://gitlab.com/AOMediaCodec/SVT-AV1/-/blob/master/Docs/CommonQuestions.md