Add a compute job specification
This adds a script which calls DataLad to provision the dataset, run
the conversion script in a container, and push the results back to the
original store.

Relying on environment variables will make it slightly easier to write
the submit file, because there will be only one (numeric) parameter to
be varied, and the others will be supplied as environment.

We may wish to place a lockfile in a clone of the dataset (especially
if we include submit script too), and it can be added to gitignore.