Make the threads used for AV1 encoding configurable via env
The number of threads to use while encoding is made configurable by
setting the environment variable n_threads or OMP_NUM_THREADS. If
neither is set, it will be set to 0, which means using all
available. Given that the number of threads is an execution detail, an
environment variable seems to be a better fit than an input parameter.

The OMP_NUM_THREADS variable is chosen because it is one of the
variables set by condor by default (to the number of CPU cores
provisioned to the job). This means that when processing on a cluster
with Condor no explicit specification is needed, and the thread limit
will be set correctly. Likewise, if computing locally (likely in the
absence of these variables), it makes sense to default to unlimited.

https://htcondor.readthedocs.io/en/latest/users-manual/env-of-job.html