Americanize the singularity image
Mkvpropedit tries to be locale-aware when reading and writing
strings. The bookworm-slim Debian base image comes with no compiled
locale, which causes mkvpropedit to error with the following error
when the container is ran without clearenv:

terminate called after throwing an instance of 'std::runtime_error'
  what():  locale::facet::_S_create_c_locale name not valid

As a one way of dealing with these issues, we install locales package
and configure the container with the en_US.UTF-8 locale. This comes at
a cost of about 5 MB added to the size of the image.