Adjust timestamps a little
Shifted by a few seconds here and there, mostly to reduce the time
with the screen displayed during breaks when the chairperson starts
talking but does not yet introduce the speaker. Two changes stand out:
day 2 part 1 (introduction) now has first 5 minutes removed (which
only showed extending the countdown and waiting for it to run down)
and day 2 part 10 (discussion) starts a bit later but also ends a bit
later because part of the last answer had been previously cut.